// packages/hertzflow-sdk-v2/src/types/trade.ts
export enum ExternalSwapAggregator {
  OpenOcean = "openOcean",
}